refactor: group priv-esc use-cases into usecases/priv_esc/
Move the privilege-escalation modules out of the flat usecases/ namespace
into a dedicated priv_esc/ subpackage, matching the existing web/, web_api/
and ad/ convention. Pure relocation + import rewiring, no behavior change:
use-cases register by class name via @use_case, so CLI names, benchmark
aliases and packaging entry points are unaffected.

- move linux_privesc, minimal_linux_privesc,
  minimal_linux_privesc_tool_calling and windows_privesc into priv_esc/
- rename the shared base _privesc_common.py -> priv_esc/_base.py
  (matches web/_base.py); subclasses import it relatively
- add priv_esc/__init__.py re-exporting the four modules so their
  @use_case decorators still run on package import
- collapse the four priv-esc star-imports in usecases/__init__.py into
  `from .priv_esc import *`
- keep call_usecase_from_usecase.py in place, repoint its PrivEscLinux import
- update the two priv-esc test modules to the new import paths
