+++ src/hackingBuddyGPT/usecases/__init__.py
@@ -1,8 +1,5 @@
 from .web import *
 from .web_api import *
-from .minimal_linux_privesc import *
-from .windows_privesc import *
-from .minimal_linux_privesc_tool_calling import *
+from .priv_esc import *
 from .call_usecase_from_usecase import *
-from .linux_privesc import *
 from .ad import *
