+++ src/hackingBuddyGPT/usecases/priv_esc/__init__.py
@@ -1,0 +1,4 @@
+from .minimal_linux_privesc import *
+from .windows_privesc import *
+from .minimal_linux_privesc_tool_calling import *
+from .linux_privesc import *
