+++ src/hackingBuddyGPT/usecases/priv_esc/_base.py
